// components/radio/radioTargets.enum.ts
export enum RadioTargets {
  root = 'root',
  label = 'label',
  subLabel = 'subLabel',
  tooltipButton = 'tooltipButton',
  tooltipPopover = 'tooltipPopover',
  coreRadioRoot = 'coreRadioRoot',
  coreRadioOptionLabel = 'coreRadioOptionLabel',
  coreRadioOptionInput = 'coreRadioOptionInput',
  readOnly = 'readOnly',
  errorsRoot = 'errorsRoot',
  errorsUnorderedList = 'errorsUnorderedList',
  errorsUnorderedListItem = 'errorsUnorderedListItem',
  errorsFieldError = 'errorsFieldError',
}